AirPressure and rain support for Surfvind

+++ Web/SurfVind_web/Surfvind/SurfvindDataConnection.cs
@@ -310,7 +310,7 @@ namespace Surfvind_2011
                     {
                         try
                         {
-                           result = decimal.Parse(reader["sum(rainFall)"].ToString());
+                           result = int.Parse(reader["sum(rainFall)"].ToString());
                         }
                         catch (Exception ee) { }
                     }
